pyproject.toml
[project]
version = "0.1.0"
name = "WhatShouldWePlay"
maintainers = [
  { name = "Somokai" },
  { name = "Javagedes" },
]
description = "A Bot to help your discord chat decide what game to play!"
readme = { file = "README.md", content-type = "text/markdown" }

dependencies = [
  "discord.py == 2.4.0",
  "pony==0.7.19",
  "python-dotenv == 1.0.1",
  "legacy-cgi == 2.6.2; python_version >= '3.13'"
]

[project.optional-dependencies]
dev = [
  "ruff == 0.8.6",
  "pytest == 8.3.4",
  "coverage == 7.6.10", 
  "dpytest @ git+https://github.com/Javagedes/dpytest.git@master",
  "requests==2.32.2"
]

[tool.coverage.run]
include = ["src"]

[tool.ruff]
line-length = 120

[lint]
ignore = ["ANN101"]
select = [
  "E", # Pycodestyle errors
  "W", # Pycodestyle warnings
  "F", # PyFlakes
  "D", # PyDocStyle
  "I", # iSort
  "PIE", # flake8
  "ANN", # function annotations
]

[lint.pydocstyle]
convention = "google"

[lint.flake8.annotations]
allow-star-arg-any = true

[tool.pytest.ini_options]
testpaths = [
  "tests"
]
